refactor: split Workbench.tsx into MergeStackModal and SlotVariablesSection components with JSDoc blocks
